+++ src/main/java/agents/Person.java
@@ -504,7 +504,7 @@ public class Person extends EconomicAgent {
      * how "far" sales inventory are from target.
      */
     @Override
-    public int estimateSupplyGap(GoodType type) {
+    public float estimateSupplyGap(GoodType type) {
         return 0;
     }
