+++ src/main/java/agents/firm/sales/pricing/MarkupFollower.java
@@ -87,7 +87,7 @@ public class MarkupFollower implements AskPricingStrategy {
      * All inventory is unwanted
      */
     @Override
-    public int estimateSupplyGap() {
+    public float estimateSupplyGap() {
         return sales.getHowManyToSell();
     }
 }
